Fix dependency cycle for commonActivityCodes
